🗺️ Atlas: Resolve plugin dependency cycle
