+++ autumn/src/lib.rs
@@ -101,12 +101,11 @@ pub use crate::i18n::t;
 pub mod mail;
 #[cfg(feature = "db")]
 pub mod migrate;
-pub mod plugin;
 pub mod plugin_conformance;
 pub mod probe;
 #[cfg(feature = "system-info")]
 pub mod system_info;
-pub use plugin::{Plugin, Plugins};
+pub use app::{Plugin, Plugins};
 
 pub mod route_listing;
